+++ pkg/apis/appstacks/v1beta1/runtimecomponent_types.go
@@ -68,9 +68,12 @@ type RuntimeComponentAutoScaling struct {
 type RuntimeComponentService struct {
 	Type *corev1.ServiceType `json:"type,omitempty"`
 
-	// +kubebuilder:validation:Maximum=65536
+	// +kubebuilder:validation:Maximum=65535
 	// +kubebuilder:validation:Minimum=1
 	Port int32 `json:"port,omitempty"`
+	// +kubebuilder:validation:Maximum=65535
+	// +kubebuilder:validation:Minimum=1
+	TargetPort *int32 `json:"targetPort,omitempty"`
 
 	PortName string `json:"portName,omitempty"`
 
@@ -412,6 +415,15 @@ func (s *RuntimeComponentService) GetAnnotations() map[string]string {
 // GetPort returns service port
 func (s *RuntimeComponentService) GetPort() int32 {
 	return s.Port
+}
+
+// GetTargetPort returns the internal target port for containers
+func (s *RuntimeComponentService) GetTargetPort() *int32 {
+	if s.TargetPort == nil {
+		return nil
+	}
+
+	return s.TargetPort
 }
 
 // GetPortName returns name of service port